Create the laser module
