+++ src/main.rs
@@ -2,38 +2,10 @@ use std::f32::consts::FRAC_PI_2;
 
 use macroquad::prelude::*;
 
-use crate::ship::SHIP_RADIUS;
+use crate::{laser::Laser, ship::SHIP_RADIUS};
 
+mod laser;
 mod ship;
-
-const LASER_VELOCITY: f32 = 200.0;
-
-#[derive(Default, Clone)]
-struct Laser {
-    position: Vec2,
-    speed: Vec2,
-    rotation_rad: f32,
-    active: bool,
-}
-
-impl Laser {
-    fn draw(&self) {
-        static V1: Vec2 = vec2(0.0, -2.0);
-        static V2: Vec2 = vec2(0.0, 2.0);
-
-        let rot_vec = Vec2::from_angle(self.rotation_rad);
-
-        let v1 = rot_vec.rotate(V1) + self.position;
-        let v2 = rot_vec.rotate(V2) + self.position;
-
-        draw_line(v1.x, v1.y, v2.x, v2.y, 1.0, RED);
-    }
-
-    fn update(&mut self, dt: f32) {
-        let direction = Vec2::from_angle(-FRAC_PI_2 + self.rotation_rad);
-        self.position += self.speed * dt + direction * LASER_VELOCITY * dt;
-    }
-}
 
 #[macroquad::main("TheGame")]
 async fn main() {
